1.11.3

+++ src/main/java/com/redsmods/sound_physics_perfected/integration/SVC.java
@@ -98,7 +98,7 @@ public class SVC implements VoicechatPlugin {
         Vec3 listenerPos = mc.player.position();
 
         // Calculate occlusion
-        float occlusion = Math.pow(Config.getInstance().permeationAbsorption,countBlocksBetween(mc.level, listenerPos, sourcePos,mc.player));
+        float occlusion = (float) Math.pow(Config.getInstance().permeationAbsorption,countBlocksBetween(mc.level, listenerPos, sourcePos,mc.player));
 
         // Apply your OpenAL effects using the source ID
         applyOpenALEffects(openALSource, occlusion);
